refactor(tauri.js): remove create command also fixes window title not setting properly because window config is an array
